// frontend/src/types/bowl.ts
export type bowl = {
    bowlerID: number;
    bowlerFirstName: string;
    bowlerMiddleInit: string;
    bowlerLastName: string;
    teamID: number;
    teamName: string;
    bowlerAddress: string;
    bowlerCity: string;
    bowlerState: string;
    bowlerZip: number;
    bowlerPhoneNumber: string;
};